+++ plugins/paintops/colorsmudge/KisColorSmudgeStrategyBase.h
@@ -22,23 +22,23 @@ public:
     {
         virtual ~DabColoringStrategy() = default;
         virtual bool supportsFusedDullingBlending() const = 0;
-        virtual void blendInColorRate(const KoColor &paintColor, const KoCompositeOp *colorRateOp, quint8 colorRateOpacity,
+        virtual void blendInColorRate(const KoColor &paintColor, const KoCompositeOp *colorRateOp, qreal colorRateOpacity,
                                       KisFixedPaintDeviceSP dstDevice, const QRect &dstRect) const = 0;
         virtual void blendInFusedBackgroundAndColorRateWithDulling(KisFixedPaintDeviceSP dst, KisColorSmudgeSourceSP src,
                                                                    const QRect &dstRect,
                                                                    const KoColor &preparedDullingColor,
                                                                    const KoCompositeOp *smearOp,
-                                                                   const quint8 smudgeRateOpacity,
+                                                                   const qreal smudgeRateOpacity,
                                                                    const KoColor &paintColor,
                                                                    const KoCompositeOp *colorRateOp,
-                                                                   const quint8 colorRateOpacity) const = 0;
+                                                                   const qreal colorRateOpacity) const = 0;
     };
 
     struct DabColoringStrategyMask : public DabColoringStrategy
     {
         bool supportsFusedDullingBlending() const override;
 
-        void blendInColorRate(const KoColor &paintColor, const KoCompositeOp *colorRateOp, quint8 colorRateOpacity,
+        void blendInColorRate(const KoColor &paintColor, const KoCompositeOp *colorRateOp, qreal colorRateOpacity,
                               KisFixedPaintDeviceSP dstDevice, const QRect &dstRect) const override;
 
         void blendInFusedBackgroundAndColorRateWithDulling(KisFixedPaintDeviceSP dst,
@@ -46,17 +46,17 @@ public:
                                                            const QRect &dstRect,
                                                            const KoColor &preparedDullingColor,
                                                            const KoCompositeOp *smearOp,
-                                                           const quint8 smudgeRateOpacity,
+                                                           const qreal smudgeRateOpacity,
                                                            const KoColor &paintColor,
                                                            const KoCompositeOp *colorRateOp,
-                                                           const quint8 colorRateOpacity) const override;
+                                                           const qreal colorRateOpacity) const override;
     };
 
     struct DabColoringStrategyStamp : public DabColoringStrategy
     {
         void setStampDab(KisFixedPaintDeviceSP device);
 
-        void blendInColorRate(const KoColor &paintColor, const KoCompositeOp *colorRateOp, quint8 colorRateOpacity,
+        void blendInColorRate(const KoColor &paintColor, const KoCompositeOp *colorRateOp, qreal colorRateOpacity,
                               KisFixedPaintDeviceSP dstDevice, const QRect &dstRect) const override;
 
         bool supportsFusedDullingBlending() const override;
@@ -66,10 +66,10 @@ public:
                                                            const QRect &dstRect,
                                                            const KoColor &preparedDullingColor,
                                                            const KoCompositeOp *smearOp,
-                                                           const quint8 smudgeRateOpacity,
+                                                           const qreal smudgeRateOpacity,
                                                            const KoColor &paintColor,
                                                            const KoCompositeOp *colorRateOp,
-                                                           const quint8 colorRateOpacity) const override;
+                                                           const qreal colorRateOpacity) const override;
 
     private:
         KisFixedPaintDeviceSP m_origDab;
@@ -91,13 +91,13 @@ public:
 
     virtual QString finalCompositeOp(bool smearAlpha) const;
 
-    virtual quint8 finalPainterOpacity(qreal opacity, qreal smudgeRateValue);
+    virtual qreal finalPainterOpacity(qreal opacity, qreal smudgeRateValue);
 
-    virtual quint8 colorRateOpacity(qreal opacity, qreal smudgeRateValue, qreal colorRateValue, qreal maxPossibleSmudgeRateValue);
+    virtual qreal colorRateOpacity(qreal opacity, qreal smudgeRateValue, qreal colorRateValue, qreal maxPossibleSmudgeRateValue);
 
-    virtual quint8 dullingRateOpacity(qreal opacity, qreal smudgeRateValue);
+    virtual qreal dullingRateOpacity(qreal opacity, qreal smudgeRateValue);
 
-    virtual quint8 smearRateOpacity(qreal opacity, qreal smudgeRateValue);
+    virtual qreal smearRateOpacity(qreal opacity, qreal smudgeRateValue);
 
     virtual void sampleDullingColor(const QRect &srcRect, qreal sampleRadiusValue, KisColorSmudgeSourceSP sourceDevice,
                                     KisFixedPaintDeviceSP tempFixedDevice, KisFixedPaintDeviceSP maskDab,
@@ -109,10 +109,10 @@ public:
                     qreal maxPossibleSmudgeRateValue, qreal colorRateValue, qreal smudgeRadiusValue);
 
     void blendInBackgroundWithSmearing(KisFixedPaintDeviceSP dst, KisColorSmudgeSourceSP src, const QRect &srcRect,
-                                       const QRect &dstRect, const quint8 smudgeRateOpacity);
+                                       const QRect &dstRect, const qreal smudgeRateOpacity);
 
     void blendInBackgroundWithDulling(KisFixedPaintDeviceSP dst, KisColorSmudgeSourceSP src, const QRect &dstRect,
-                                      const KoColor &preparedDullingColor, const quint8 smudgeRateOpacity);
+                                      const KoColor &preparedDullingColor, const qreal smudgeRateOpacity);
 
 protected:
     const KoCompositeOp * m_colorRateOp {nullptr};
